Refactor Alert
