+++ app/client/components/CustomToast.tsx
@@ -1,11 +1,11 @@
 import { Transition } from "@headlessui/react";
 import React from "react";
-import { Alert } from "./Alert";
+import { Alert, AlertType } from "./alert/Alert";
 
 interface CustomToastProps {
   isToastVisible: boolean;
   message: string;
-  levelTag: "error" | "warning" | "info" | "success" | "debug";
+  levelTag: AlertType;
 }
 
 export const CustomToast = ({
